[FEATURE] Add extension to fix TypoScript Rendering with RealURL
Whenever the extension »TypoScript Rendering« is used RealURL
tries to rewrite some AJAX URLs which should not be rewritten.

To avoid this a hook will look for »typoscript rendering« parameters
in the URL and break the rewrite action.

The hook is part of a bugfix extension → so if we need Ajax
then require the bugfix extension and it will then load
realurl & typoscript_rendering automatically.

Refs #45156